Fix the javacomplete issue.
For years I could not use javacomplete because for some reason it openend the
quickfix window when the omnifunc fuction failed for some reason making it
impossible to edit efficiently in Java. After so many years disabling the javacomplete
plugin I just found it was all my fault.

I had a hook that automatically opens and focuses the quickfix window if there are errors
present. So when the omnifunc failed with some error the quickfix was opened
and focused breaking my typing flow.

Removing those hooks now I can enjoy the virtues of Java autocompletion and
even better using the improved version from Shougo.