Replace gtags with gutentags
Unfortunatelly current state of global (gtags) is terrible in Ubuntu. It only
works for the included parsers and the pygments plugin that enables other
languages requires python2.7 that no longer is default in Ubuntu 16.04.

Gutentags with Universal Ctags works mostly OK for go, ruby, python, js and java
that are the main languages I use now.